refactor: SseEmittersService를 global.sse 패키지로 이동
