+++ src/main/java/com/example/onlyone/global/sse/SseStreamController.java
@@ -1,6 +1,5 @@
 package com.example.onlyone.global.sse;
 
-import com.example.onlyone.domain.notification.service.SseEmittersService;
 import com.example.onlyone.domain.user.entity.User;
 import com.example.onlyone.domain.user.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
